# apps/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "오류 식별 코드",
            "example": "QUESTION_SET_NOT_FOUND",
            "title": "Code",
            "type": "string"
          },
          "fieldErrors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "필드별 유효성 검증 오류 맵 (선택적)",
            "example": {},
            "title": "Fielderrors",
            "type": "object"
          },
          "message": {
            "description": "오류 상세 메시지",
            "example": "요청한 버전의 질문 세트를 찾을 수 없습니다.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "오류 세부 정보 객체"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ReviewerContextResponse": {
        "properties": {
          "demo": {
            "const": true,
            "default": true,
            "title": "Demo",
            "type": "boolean"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "role": {
            "enum": [
              "A",
              "B"
            ],
            "title": "Role",
            "type": "string"
          },
          "roomCode": {
            "title": "Roomcode",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "role",
          "roomCode",
          "expiresAt"
        ],
        "title": "ReviewerContextResponse",
        "type": "object"
      },
      "ReviewerLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "roomCode": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roomcode"
          },
          "username": {
            "enum": [
              "judge-a",
              "judge-b"
            ],
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "ReviewerLoginRequest",
        "type": "object"
      },
      "ReviewerResetRequest": {
        "additionalProperties": false,
        "properties": {
          "confirm": {
            "title": "Confirm",
            "type": "boolean"
          }
        },
        "required": [
          "confirm"
        ],
        "title": "ReviewerResetRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "accountAuth": {
        "in": "cookie",
        "name": "mrs_account",
        "type": "apiKey"
      },
      "cookieAuth": {
        "description": "참여자 세션 인증을 위한 HttpOnly 쿠키",
        "in": "cookie",
        "name": "mrs_participant",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "신혼부부를 위한 맞춤형 재무 진단, 세션 관리 및 라이트/딥 진단 연산 API 모듈입니다. 모든 질문, 선택지, 결과 해설을 한국어로 지원합니다.",
    "title": "미리살림 백엔드 API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/reviewer/context": {
      "get": {
        "operationId": "reviewer_context_api_v1_auth_reviewer_context_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewerContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "입력값 검증 실패 (Validation Error)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "accountAuth": []
          }
        ],
        "summary": "Reviewer Context",
        "tags": [
          "reviewer"
        ]
      }
    },
    "/api/v1/auth/reviewer/login": {
      "post": {
        "operationId": "reviewer_login_api_v1_auth_reviewer_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewerLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewerContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "입력값 검증 실패 (Validation Error)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Reviewer Login",
        "tags": [
          "reviewer"
        ]
      }
    },
    "/api/v1/auth/reviewer/reset": {
      "post": {
        "operationId": "reviewer_reset_api_v1_auth_reviewer_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewerResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewerContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "입력값 검증 실패 (Validation Error)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "accountAuth": []
          }
        ],
        "summary": "Reviewer Reset",
        "tags": [
          "reviewer"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "기본 서버 (Default)",
      "url": "/"
    }
  ]
}